+++ src/main/java/com/zylitics/api/controllers/BuildController.java
@@ -54,6 +54,8 @@ public class BuildController extends AbstractController {
   
   private final APIDefaultsProvider apiDefaultsProvider;
   
+  private final TestProvider testProvider;
+  
   private final VMService vmService;
   
   private final RunnerService runnerService;
@@ -67,6 +69,7 @@ public class BuildController extends AbstractController {
                          UserPlanProvider userPlanProvider,
                          BrowserProvider browserProvider,
                          APIDefaultsProvider apiDefaultsProvider,
+                         TestProvider testProvider,
                          APICoreProperties apiCoreProperties,
                          BuildCompletionEmailHandler buildCompletionEmailHandler,
                          Common common) {
@@ -77,6 +80,7 @@ public class BuildController extends AbstractController {
     this.userPlanProvider = userPlanProvider;
     this.browserProvider = browserProvider;
     this.apiDefaultsProvider = apiDefaultsProvider;
+    this.testProvider = testProvider;
     this.apiCoreProperties = apiCoreProperties;
     this.buildCompletionEmailHandler = buildCompletionEmailHandler;
     this.common = common;
@@ -291,6 +295,9 @@ public class BuildController extends AbstractController {
               .orElseThrow(() -> new RuntimeException("Couldn't fetch build " + buildId));
           res.setStatus(build.getFinalStatus().toString().toLowerCase(Locale.US));
           res.setError(build.getError());
+          if (config.isIncludeDetailedResultInResponse()) {
+            res.setTestDetails(testProvider.getAllCompletedTestDetail(buildId));
+          }
         } else {
           res.setStatus(TestStatus.RUNNING.toString().toLowerCase(Locale.US));
         }